refactor: split Sample and MiniBatch into separate modules
- Moved `Sample` from `src/lib.rs` to `src/sample.rs`
- Moved `MiniBatch` from `src/lib.rs` to `src/minibatch.rs`
- Updated `src/lib.rs` to expose both modules
- Left unit tests inline within each module for simplicity
